% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.R
\name{dataRange}
\alias{dataRange}
\title{dataRange function}
\usage{
dataRange(X)
}
\arguments{
\item{X}{A matrix or a data frame of training data \strong{without} y column.}
}
\value{
A data frame with two columns \code{min max} contains the range of each x variable.
}
\description{
Return the range of X for each variable.
}
\details{
Return the range of X for each variable, which is a part of \code{param}. Can be used for passing \code{param} to \code{Elem, ORT or ORF}.
}
\examples{
dataRange(iris[, 1:4])
dataRange(mtcars)
}
\author{
Quan Gu
}
