% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORTree.R
\docType{data}
\name{Elem}
\alias{Elem}
\title{Create a Node Element Object}
\format{\code{\link{R6Class}} object.}
\usage{
Elem$new(param, splitInd = -1, splitVal = 0, numSamplesSeen = 0)
}
\arguments{
\item{param}{A list which usually has names of \code{minSamples, minGain, numClasses, x.rng, etc.}}

\item{splitInd}{A integer indicates which x variable is used to split on current node. Default NULL.}

\item{splitVal}{A numeric indicates what value of the chosen x variable is used to split on current node. Means: \code{x[splitInd] < splitVal}. Default NULL.}

\item{numSamplesSeen}{A integer indicates how many samples have come in current node for updating the ORT tree. Default 0.}
}
\value{
Object of \code{\link{R6Class}}, Object of \code{Node Element}.
}
\description{
Elem is a class of R6. It is the element of a node in a Tree or ORT object, storing some important information of the node/Tree.
}
\details{
See details in description of each field or method.
}
\section{Fields}{

\describe{
\item{\code{x.rng}}{
A data frame which indicates the range of every x variable in training data.
It must be a shape of \code{n*2} which n is the number of x variables, i.e. \code{x.dim}.
And the first collumn must be the minimal values of x and the second as maximum.
You can generate it via \code{OnlineRandomForest::dataRange()} for convenience.
}
\item{\code{x.dim}}{Number of x variables.}
\item{\code{tests}}{A list of \code{SuffStats}. Candicate splits of current node.}
\item{\code{numTests}}{A part of \code{param} indicates the number of \code{SuffStats} in tests}
\item{\code{stats}}{A \code{SuffStats} object. The real splits of current node.}
}
}

\section{Methods}{

\describe{
\item{\code{toString()}}{Print elem itself or the split of condition on current node.}
\item{\code{pred()}}{Return the prediction of current leaf node. Will return an \strong{integer} for classification or an \strong{numeric} for regression.}
\item{\code{update(x, y)}}{
When a sample comes in current node, update ORT with the sample's x variables and y value. \cr
\itemize{
\item x - The x variables of a sample. Note it is an numeric vector other than a scalar.
\item y - The y value of a sample.
}
}
\item{\code{updateSplit(xvar.index, xvar.value)}}{
Replace the fields \strong{splitInd} and \strong{splitVal} of current node with xvar.index and xvar.value.
}
\item{...}{Other functions, usually not used.}
}
}

\examples{
x.rng <- data.frame(min = c(1,3,5,7), max = c(2,4,6,8), row.names = paste0("x",1:4))
param <- list('minSamples'= 5, 'minGain'= 0.1, 'numClasses'= 3, 'x.rng'= x.rng)
elm <- Elem$new(param)
length(elm$tests)
elm$tests[c(1,8)]
elm$updateSplit(3, 5.2)
elm$toString()
elm$update(c(1.1, 3.2, 5.3, 7.4), 1)

}
\author{
Quan Gu
}
\keyword{datasets}
