% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORTree.R
\docType{data}
\name{SuffStats}
\alias{SuffStats}
\title{Create a SuffStats Object}
\format{\code{\link{R6Class}} object.}
\usage{
SuffStats$new(numClasses = 0)
}
\arguments{
\item{numClasses}{A nonnegative integer indicates how many classes when solve a classifation problem. Default 0 for regression. If numClasses > 0, then do classifation.}
}
\value{
Object of \code{\link{R6Class}}, Object of \code{SuffStats}.
}
\description{
SuffStats is a class of R6. It has provide some important statistics(fields or functions) inside the element of every (leaf) node in an \code{\link{ORT}} object.
}
\details{
Note that SuffStats may only be seen in leaf nodes of an ORT tree. See details in description of each field or method.
}
\section{Fields}{

\describe{
\item{\code{node.n}}{Number of samples under current node. Default 0.}
\item{\code{classify}}{TRUE for classification and FALSE for Regression.}
\item{\code{node.counts}}{If classification, counts of each y value in current node.}
\item{\code{node.sum}}{Sum of the y value in current node.}
\item{\code{node.square.sum}}{Sum of the y's square in current node.}
}
}

\section{Methods}{

\describe{
\item{\code{update(y)}}{
When a sample comes in current node, update ORT with the sample's y value. \cr
\itemize{
\item y - The y value of a sample.
}
}
\item{\code{reset()}}{Reset all fields to NULL. Currently not used.}
\item{\code{isLeaf()}}{TRUE if current node is a leaf node otherwise FALSE.}
\item{\code{pred()}}{Return the prediction of current leaf node. Will return an \strong{integer} for classification or an \strong{numeric} for regression.}
\item{\code{impurity()}}{Return the impurity of current node. Computed via information entropy.}
}
}

\examples{
# classification
sa <- SuffStats$new(numClasses = 3)
sa$update(2);sa$update(1);sa$update(2)
sa$pred()
sa$node.counts
sa$node.n
sa$impurity()

# Regression
sb <- SuffStats$new(numClasses = 0)
sb$update(1);sb$update(2);sb$update(3)
sb$pred()
sb$node.sum;sb$node.square.sum
sb$impurity()

}
\author{
Quan Gu
}
\keyword{datasets}
