% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.R
\docType{data}
\name{Test}
\alias{Test}
\title{Create a Candidate Split Object}
\format{\code{\link{R6Class}} object.}
\usage{
Test$new(xvar.index, xvar.value, numClasses = 0)
}
\arguments{
\item{xvar.index}{A integer indicates which x variable is used to split.}

\item{xvar.value}{A numeric indicates what value of the chosen x variable is used to split. Means: \code{x[xvar.index] < xvar.value}.}

\item{numClasses}{A nonnegative integer indicates how many classes when solve a classifation problem. Default 0 for regression. If numClasses > 0, then do classifation.}
}
\value{
Object of \code{\link{R6Class}}, Object of \code{Candidate Split}.
}
\description{
Test is a class of R6. It is the candicate split of a node. Named by \emph{Amir Saffari} 's paper.
}
\details{
Note that Test may only be seen in leaf nodes of an ORT tree. See details in description of each field or method.
}
\section{Fields}{

\describe{
\item{\code{statsL}}{A SuffStats object in left hand of a split.}
\item{\code{statsR}}{A SuffStats object in right hand of a split.}
}
}

\section{Methods}{

\describe{
\item{\code{update(x, y)}}{
When a sample comes in current node, update ORT with the sample's x variables and y value. \cr
\itemize{
\item x - The x variables of a sample. Note it is an numeric vector other than a scalar.
\item y - The y value of a sample.
}
}
}
}

\examples{
t1 <- Test$new(3, 2.2, 0)
t1$update(c(0,0,2,4), 1.2)
t1$update(c(0,0,1,4), 1.4)
t1$update(c(1,1,3,3), 2.7)
t1$statsL$pred()
t1$statsR$pred()

}
\author{
Quan Gu
}
\keyword{datasets}
